Added Support For Attributes by Resource ID
Added Support For Attributes by Resource ID. This was necessary for certain ManifestFiles that had an empty attributeName, and only used the attributeID.
Also, fixed bug that had the littleEndianWord function return the unsigned value rather than the signed value in 64 bit systems. Thus, now instead of checking for 0xffffffff, we check for -1, which works on all systems.